Adding modules and deployment changes
