+++ lib/legoTech_gem.rb
@@ -1,0 +1,10 @@
+require_relative "ui/actions/Action"
+
+require_relative "ui/driver/Driver"
+
+require_relative "ui/test-cases/TestCase"
+
+require_relative "ui/test-suites/TestSuite"
+
+require_relative "ui/enums/Field"
+require_relative "ui/enums/Identifier"
