--- lib/main/ruby/SeleniumRuby.rb
@@ -1,13 +1,0 @@
-require "actions/Action"
-
-require "driver/Driver"
-
-require "test-cases/TestCase"
-
-require "test-suites/TestSuite"
-
-require "enums/Field"
-require "enums/Identifier"
-
-module SeleniumRuby
-end
