--- lib/main/ruby/enums/Field.rb
@@ -1,6 +1,0 @@
-module Fields
-    RADIO_BUTTON      = 1
-    CHECK_BOX         = 2
-    INPUT             = 3
-    DROPDOWN_LIST     = 4
-end
