--- lib/main/ruby/enums/Identifier.rb
@@ -1,5 +1,0 @@
-module Identifier
-    CSS    = 1
-    NAME   = 2
-    ID     = 3
-end
