--- lib/main/ruby/test-cases/TestCase.rb
@@ -1,9 +1,0 @@
-class TestCase
-  def initialize(name)
-    if name.nil? or name.empty?
-      raise "All TestCasses must have a name associated to them"
-    end
-
-    @name = name
-  end
-end
