+++ lib/ui/enums/Field.rb
@@ -1,0 +1,8 @@
+module LegoTechSelenium
+  module Fields
+    RADIO_BUTTON = 1
+    CHECK_BOX = 2
+    INPUT = 3
+    DROPDOWN_LIST = 4
+  end
+end
