+++ lib/ui/enums/Identifier.rb
@@ -1,0 +1,7 @@
+module LegoTechSelenium
+  module Identifier
+    CSS = 1
+    NAME = 2
+    ID = 3
+  end
+end
