+++ test/ui/test-suites/TestSuiteTest.rb
@@ -1,10 +1,10 @@
 require "test/unit"
-require_relative "../../../main/ruby/test-suites/TestSuite"
-require_relative "../../../main/ruby/driver/Driver"
+require_relative "../../../lib/ui/test-suites/TestSuite"
+require_relative "../../../lib/ui/driver/Driver"
 
 class TestSuiteTest < Test::Unit::TestCase
   def test_suites_basic
-    testSuites = TestSuite::Builder.new("My Sweet", Driver.new(""))
+    testSuites = LegoTechSelenium::TestSuite::Builder.new("My Sweet", LegoTechSelenium::Driver.new(""))
       .addTestCase("Test Case 1")
       .addTestCase("Test Case 2")
       .addTestCase("Test Case 3")
